Add bippy to the project
