+++ package.json
@@ -123,6 +123,7 @@
 		"@vitest/coverage-v8": "^2.1.5",
 		"@vitest/ui": "^2.1.5",
 		"autoprefixer": "^10.4.20",
+		"bippy": "^0.2.2",
 		"glob": "^11.0.0",
 		"happy-dom": "^15.7.4",
 		"jest-preview": "^0.3.1",
